+++ src/compute-engine/interval/index.ts
@@ -165,6 +165,9 @@ export {
   gammaln,
   factorial,
   factorial2,
+  binomial,
+  gcd,
+  lcm,
 } from './elementary';
 
 // Trigonometric functions